atualização estados detailList
