+++ pokedex7/src/pages/DetailList/DetailList.js
@@ -1,7 +1,8 @@
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-import { BASE_URL } from '../../constants/baseUrl';
-import { Link } from 'react-router-dom';
+import React, { useContext } from "react";
+// import axios from 'axios';
+// import { BASE_URL } from '../../constants/baseUrl';
+import { Link } from "react-router-dom";
+import { PokeContext } from "../../Components/Context/PokeContext";
 import {
   Container,
   FrontBack,
@@ -9,27 +10,29 @@ import {
   StatsCard,
   NameStat,
   BaseStat,
-} from './styles';
+} from "./styles";
 
 export const DetailList = () => {
-  const [stats, setStats] = useState([]);
-  const [image, setImage] = useState({});
-  const [moves, setMoves] = useState([]);
+  // const [stats, setStats] = useState([]);
+  // const [image, setImage] = useState({});
+  // const [moves, setMoves] = useState([]);
 
-  const getDetails = () => {
-    axios.get(`${BASE_URL}/pokemon/bulbasaur`).then((res) => {
-      setStats(res.data.stats);
-      setImage(res.data.sprites);
-      setMoves(res.data.moves);
-      //   console.log(res.data.moves.length);
-      //   console.log(res.data.moves);
-    });
-  };
-  //   console.log(image);
+  const { stats, moves, image } = useContext(PokeContext);
 
-  useEffect(() => {
-    getDetails();
-  }, []);
+  // const getDetails = () => {
+  //   axios.get(`${BASE_URL}/pokemon/bulbasaur`).then((res) => {
+  //     setStats(res.data.stats);
+  //     setImage(res.data.sprites);
+  //     setMoves(res.data.moves);
+  //     //   console.log(res.data.moves.length);
+  //     //   console.log(res.data.moves);
+  //   });
+  // };
+  // //   console.log(image);
+
+  // useEffect(() => {
+  //   getDetails();
+  // }, []);
 
   const renderStats = stats.map((pokemon, index) => {
     return (
@@ -40,35 +43,29 @@ export const DetailList = () => {
     );
   });
 
-  //   const renderMoves = moves.map((item) => {
-  //     return <p>{item.move.name}</p>;
-  //   });
-
-  const renderMoves = (array) => {
-    for (let i = 0; i < 4; i++) {
-      console.log(array[i].move.name);
-    }
-  };
+    // const renderMoves = moves.map((item) => {
+    //   return <p>{item.move.name}</p>;
+    // });
 
   return (
     <div>
-      <Link to={'/'}>
+      <Link to={"/"}>
         <button>Voltar</button>
       </Link>
       <p>Aqui vao aparecer os detalhes de pokemon em questao</p>
       <Container>
         <FrontBack>
           <div>
-            <img src={image.front_default} />
+            <img src={image.front_default} alt="" />
           </div>
           <div>
-            <img src={image.back_default} />
+            <img src={image.back_default} alt="" />
           </div>
         </FrontBack>
         <Stats>{renderStats}</Stats>
         <div>
           {/* <img src={image.other.dream_world.front_default} /> */}
-          <p>{renderMoves(moves)}</p>
+          {/* <p>{renderMoves(moves)}</p> */}
         </div>
       </Container>
       {/* <div>
